Downgrade React to v18, update dependencies, and refactor formatters
